E-commerce_cloud_native/requirements.txt
streamlit
azure-storege-blob
pymssql
dotenv